# Disable predicate pushdown filters in MULTITHREADED Delta Scan when reading Deletion Vectors