--- a/delta-lake/delta-33x/src/main/scala/com/nvidia/spark/rapids/delta/delta33x/GpuDelta33xParquetFileFormat.scala
+++ b/delta-lake/delta-33x/src/main/scala/com/nvidia/spark/rapids/delta/delta33x/GpuDelta33xParquetFileFormat.scala
@@ -232,7 +232,7 @@ case class GpuDelta33xParquetFileFormat(
       prepareSchema(fileScan.relation.dataSchema),
       prepareSchema(fileScan.requiredSchema),
       prepareSchema(fileScan.readPartitionSchema),
-      pushedFilters,
+      prepareFiltersForRead(pushedFilters).toArray,
       fileScan.rapidsConf,
       fileScan.allMetrics,
       useMetadataRowIndex = false,