Fixes #13513 

### Description

- When we added support for reading Deletion Vectors on Delta Lake 3.3.0 we didn't disable the predicate pushdown which caused a misalignment while reading the Delta Table. This PR fixes the data corruption that would happen if the predicates were pushed down and row groups were dropped. 
- A new integration test has been added 

Example: Assume only the first two values (-3268, -3267) are deleted 
<table>
  <tr>
    <th colspan="1"></th>
    <th colspan="2">Row group 1</th>
    <th colspan="2">Row group 2</th>
    <th colspan="2">Row group 3</th>
  </tr>
  <tr>
    <td/>
    <td>-3268</td>
    <td>-3267</td>
    <td>0</td>
    <td>1</td>
    <td>3266</td>
    <td>3267</td>
  </tr>
  <tr> 
    <td>dv</td>
    <td>true</td>
    <td>true</td>
    <td>false</td>
    <td>false</td>
    <td>false</td>
    <td>false</td>     
</table>

After dropping Row group 1 and 2, we have a misalignment and now it shows 3266 and 3267 as deleted
<table>
  <tr>
    <th colspan="1"></th>
    <th colspan="2">Row group 3</th>
  </tr>
  <tr>
    <td/>
    <td>3266</td>
    <td>3267</td>
  </tr>
  <tr> 
    <td>dv</td>
    <td>true</td>
    <td>true</td>
    <td>false</td>
    <td>false</td>
    <td>false</td>
    <td>false</td>     
</table>

#### Performance 
Performance runs against NDS SF100 dataset on workstation

  | Baseline - commit id 5d88f6a | DV MULTITHREADED READER |  
-- | -- | -- | --
  | Average | Average | Speedup
5% | 40327.2 | 29165.2 | 1.38
10% | 39161.4 | 29577 | 1.32
25% | 39043.8 | 29784.6 | 1.31
50% | 37682.4 | 28661 | 1.31
75% | 37800 | 28600 | 1.32
50% small files | 36465.2 | 28084.8 | 1.30

Query used 

```
select *
from store_sales
order by rand(12324) 
limit 1000000
```


### Checklists

- [ ] This PR has added documentation for new or modified features or behaviors.
- [x] This PR has added new tests or modified existing tests to cover new code paths.
- [x] Performance testing has been performed and its results are added in the PR description. Or, an issue has been filed with a link in the PR description.